Split out jobs handling from downsample processor class into its own class
