+++ src/main/java/com/rackspace/ceres/app/services/DownsampleProcessor.java
@@ -31,152 +31,35 @@ import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-import javax.annotation.PostConstruct;
-import javax.annotation.PreDestroy;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.Iterator;
 import java.util.List;
-import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
-import java.util.stream.IntStream;
 
-import static com.rackspace.ceres.app.utils.DateTimeUtils.*;
+import static com.rackspace.ceres.app.utils.DateTimeUtils.epochToLocalDateTime;
 
 @Service
 @Slf4j
 @Profile("downsample")
 public class DownsampleProcessor {
-
   private final DownsampleProperties properties;
-  private final DownsampleTrackingService trackingService;
-  private final DelayedTrackingService delayedTrackingService;
   private final QueryService queryService;
   private final DataWriteService dataWriteService;
   private final Timer meterTimer;
-  private final ScheduledExecutorService executor;
 
   @Autowired
   public DownsampleProcessor(DownsampleProperties properties,
-                             DownsampleTrackingService trackingService,
-                             DelayedTrackingService delayedTrackingService,
                              QueryService queryService,
                              DataWriteService dataWriteService,
-                             MeterRegistry meterRegistry,
-                             ScheduledExecutorService executor) {
+                             MeterRegistry meterRegistry) {
     this.properties = properties;
-    this.trackingService = trackingService;
-    this.delayedTrackingService = delayedTrackingService;
     this.queryService = queryService;
     this.dataWriteService = dataWriteService;
     this.meterTimer = meterRegistry.timer("downsampling.delay");
-    this.executor = executor;
   }
 
-  @PostConstruct
-  public void setupSchedulers() {
-    if (properties.getGranularities() == null ||
-        properties.getGranularities().isEmpty()) {
-      throw new IllegalStateException("Granularities are not configured!");
-    }
-    long initialDelay = properties.getInitialProcessingDelay().getSeconds();
-    executor.schedule(this::initializeRedisJobs, 1, TimeUnit.SECONDS);
-    executor.schedule(this::initializeJobs, initialDelay, TimeUnit.SECONDS);
-    executor.schedule(this::initializeDelayedTimeslotJobs, initialDelay, TimeUnit.SECONDS);
-  }
-
-  @PreDestroy
-  public void stop() {
-    executor.shutdown();
-  }
-
-  private void initializeRedisJobs() {
-    getPartitionWidths(properties.getGranularities())
-        .forEach(width -> IntStream.rangeClosed(0, properties.getPartitions() - 1)
-            .forEach((partition) -> trackingService.freeJob(partition, width).subscribe()));
-    IntStream.rangeClosed(0, properties.getPartitions() - 1)
-        .forEach((partition) -> delayedTrackingService.freeJob(partition).subscribe());
-  }
-
-  private void initializeJobs() {
-    log.info("Start downsampling jobs");
-    log.info("Downsampling configuration parameters");
-    log.info("=====================================");
-    log.info("downsample-spread-period: {}", properties.getDownsampleSpreadPeriod().getSeconds());
-    log.info("max-concurrent-downsample-hashes: {}", properties.getMaxConcurrentDownsampleHashes());
-    log.info("max-downsample-job-duration: {}", properties.getMaxDownsampleJobDuration().getSeconds());
-    log.info("downsample-delayed-timeslot-period: {}", properties.getDownsampleDelayedTimeslotPeriod().getSeconds());
-    log.info("=====================================");
-
-    getPartitionWidths(properties.getGranularities())
-        .forEach(width -> IntStream.rangeClosed(0, properties.getPartitions() - 1)
-            .forEach((partition) -> executor.schedule(() ->
-                    processJob(partition, width),
-                randomDelay(properties.getDownsampleSpreadPeriod().getSeconds()), TimeUnit.SECONDS)));
-  }
-
-  private void initializeDelayedTimeslotJobs() {
-    IntStream.rangeClosed(0, properties.getPartitions() - 1)
-        .forEach((partition) -> executor.schedule(
-            () -> processDelayedTimeslotJob(partition),
-            properties.getDownsampleDelayedTimeslotPeriod().getSeconds(), TimeUnit.SECONDS));
-  }
-
-  private void processJob(int partition, String group) {
-    log.trace("processJob {} {}", partition, group);
-    trackingService.claimJob(partition, group)
-            .flatMap(status -> status.equals("free") ?
-                processTimeSlot(partition, group).then(trackingService.freeJob(partition, group)) : Flux.empty()
-            ).subscribe();
-    executor.schedule(() -> processJob(partition, group),
-        randomDelay(properties.getDownsampleSpreadPeriod().getSeconds()), TimeUnit.SECONDS);
-  }
-
-  private void processDelayedTimeslotJob(int partition) {
-    log.trace("processDelayedTimeslotJob {}", partition);
-    delayedTrackingService.claimJob(partition)
-        .flatMap(status -> status.equals("free") ?
-            processDelayedTimeSlot(partition)
-                .then(delayedTrackingService.freeJob(partition)) : Flux.empty()
-        ).subscribe();
-    executor.schedule(() -> processDelayedTimeslotJob(partition),
-        properties.getDownsampleDelayedTimeslotPeriod().getSeconds(), TimeUnit.SECONDS);
-  }
-
-  private Flux<?> processTimeSlot(int partition, String group) {
-    log.trace("processTimeSlot {} {}", partition, group);
-    return trackingService.getTimeSlot(partition, group)
-        .flatMapMany(ts -> {
-          long timeslot = Long.parseLong(ts);
-          log.info("Got timeslot: {} {} {}", partition, group, epochToLocalDateTime(timeslot));
-          return trackingService.getDownsampleSets(timeslot, partition)
-              .name("processTimeSlot")
-              .tag("partition", String.valueOf(partition))
-              .tag("group", group)
-              .metrics()
-              .flatMap(downsampleSet -> processDownsampleSet(downsampleSet, partition, group),
-                  properties.getMaxConcurrentDownsampleHashes())
-              .then(trackingService.deleteTimeslot(partition, group, timeslot))
-              .doOnError(Throwable::printStackTrace);
-        });
-  }
-  private Flux<?> processDelayedTimeSlot(int partition) {
-    return Flux.fromIterable(getPartitionWidths(properties.getGranularities()))
-        .flatMap(group -> delayedTrackingService.getDelayedTimeSlots(partition, group)
-                .map(tsString -> DelayedTrackingService.buildDownsampleSet(partition, group, tsString))
-                .name("processDelayedTimeSlot")
-                .tag("partition", String.valueOf(partition))
-                .tag("group", group)
-                .metrics()
-                .flatMap(downsampleSet -> Mono.from(processDownsampleSet(downsampleSet, partition, group))
-                        .then(delayedTrackingService.deleteDelayedTimeslot(
-                            partition, group, DelayedTrackingService.encodeDelayedTimeslot(downsampleSet))))
-                .doOnError(Throwable::printStackTrace),
-            properties.getMaxConcurrentDownsampleHashes()
-        );
-  }
-
-  private Publisher<?> processDownsampleSet(PendingDownsampleSet pendingSet, int partition, String group) {
+  public Publisher<?> processDownsampleSet(PendingDownsampleSet pendingSet, int partition, String group) {
     log.trace("processDownsampleSet {} {} {}", pendingSet, partition, group);
     Duration downsamplingDelay = Duration.between(pendingSet.getTimeSlot(), Instant.now());
     this.meterTimer.record(downsamplingDelay.getSeconds(), TimeUnit.SECONDS);
